requirements.txt

weaviate_client==3.2.2
pandas==1.0.3
tqdm
uuid
json